feat: add function to regrid an image and header class
This commit adds a number of new features related to handling anatomical images
and "regridding" axis-aligned images.

* Port `Header` from aind-zarr-utils to this library, and rename it to
`AnatomicalHeader`. Changed the implementation details of `Header` as
testing here revealed bugs in previous implementation related to ANTs images.

* Add method `orientation_code` to `AnatomicalHeader`. `orientation_code`
returns the DICOM orientation code for the header.

* Add method `regrid_to` to `AnatomicalHeader`, which calculates the changes
required to redfine the voxel grid while maintaining the same physical points
for an anatomical image. This does not operate on the voxels themselves, and is
only concerned with the image domain.

* Add functions `regrid_axis_aligned_sitk` and `regrid_axis_aligned_ants` which
will take a sitk image or ants image, respectively, and regrid them while
maintaining the same physical interpretation and avoiding changing the voxel
values.

* Add tests for the new functions and for many of the `Header` methods which
were not tested previously.

* Add new dependency `antspyx`

* Tweak `run_linters_and_checks` to not stop on first error.